+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 0.2.3 
+* ~~Support `Reading mode`~~.  
+* Supported [markdown-it-emoji](https://github.com/markdown-it/markdown-it-emoji) and [font-awesome](https://github.com/FortAwesome/Font-Awesome).  
+* Fixed pandoc code block bug.  
+
 # 0.2.2
 * Added `enableCriticMarkupSyntax` option. enableCriticMarkupSyntax. [Syntax guide](http://criticmarkup.com/users-guide.php).  
 * Added `toc` config to front-matter for `[TOC]` and sidebar TOC [#606](https://github.com/shd101wyy/markdown-preview-enhanced/issues/606).